+++ src/components/pageComponents/for-detail/Likes.jsx
@@ -1,0 +1,9 @@
+const Likes = () => {
+    return (
+        <p className="text-2xl text-black font-bold">
+            <span className="text-2xl text-red-500 font-bold">♥</span> 5
+        </p>
+    );
+};
+
+export default Likes;
